refactor: reshape packages and refactor run functions
