--- sem/until.go
@@ -1,48 +1,0 @@
-package sem
-
-import (
-	"sync"
-	"time"
-
-	"github.com/benchttp/runner/request"
-)
-
-// RunUntil launches a goroutine to ping url as soon as a thread is
-// available and sends the results through a pipeline as they come in.
-// Returns a channel to pipeline the records.
-// The value of concurrency limits the number of concurrent threads.
-// On quit signal, waits for goroutines and closes the records channel.
-func RunUntil(quit <-chan struct{}, concurrency int, url string, timeout time.Duration) <-chan request.Record {
-	// sem is a semaphore to constrain access to at most n concurrent threads.
-	sem := make(chan int, concurrency)
-	rec := make(chan request.Record)
-
-	var wg sync.WaitGroup
-
-	acquire := func() {
-		sem <- 1
-		wg.Add(1)
-	}
-	release := func() {
-		<-sem
-		wg.Done()
-	}
-
-	go func() {
-		for {
-			select {
-			case <-quit:
-				wg.Wait()
-				close(rec)
-			default:
-			}
-			acquire()
-			go func() {
-				defer release()
-				rec <- request.Do(url, timeout)
-			}()
-		}
-	}()
-
-	return rec
-}
